Disable temporary -A on 65535
Before radare2 in aac took section->size instead of vsize when being on
virtual address. Changing that makes r2 to scan the whole 65535 maps
which is valid but introduces a huge slow down.

This should be reviewed to see how to cope with such cases. Likely scan
mmap in 99% of cases won't throw useful data so those could be avoided.